+++ main.go
@@ -21,24 +21,25 @@ import (
 )
 
 var (
-	VALID_REMOVE_ACTIONS = []string{"ignore", "delete", "rename"}
+	// Never delete rbd images
+	VALID_REMOVE_ACTIONS = []string{"ignore", "rename"}
 
 	// Plugin Option Flags
 	versionFlag        = flag.Bool("version", false, "Print version")
 	debugFlag          = flag.Bool("debug", false, "Debug output")
 	pluginName         = flag.String("name", "rbd", "Docker plugin name for use on --volume-driver option")
 	cephUser           = flag.String("user", "admin", "Ceph user")
-	cephConfigFile     = flag.String("config", "/etc/ceph/ceph.conf", "Ceph cluster config") // more likely to have config file pointing to cluster
-	cephCluster        = flag.String("cluster", "", "Ceph cluster")                          // less likely to run multiple clusters on same hardware
+	cephConfigFile     = flag.String("config", "/etc/ceph/xtao.conf", "Xtao ceph cluster config") // more likely to have config file pointing to cluster
+	cephCluster        = flag.String("cluster", "xtao", "xtao ceph cluster")                      // less likely to run multiple clusters on same hardware
 	defaultCephPool    = flag.String("pool", "rbd", "Default Ceph Pool for RBD operations")
 	pluginDir          = flag.String("plugins", "/run/docker/plugins", "Docker plugin directory for socket")
 	rootMountDir       = flag.String("mount", dkvolume.DefaultDockerRootDirectory, "Mount directory for volumes on host")
 	logDir             = flag.String("logdir", "/var/log", "Logfile directory")
-	canCreateVolumes   = flag.Bool("create", false, "Can auto Create RBD Images")
+	canCreateVolumes   = flag.Bool("create", true, "Can auto Create RBD Images")
 	defaultImageSizeMB = flag.Int("size", 20*1024, "RBD Image size to Create (in MB) (default: 20480=20GB)")
-	defaultImageFSType = flag.String("fs", "xfs", "FS type for the created RBD Image (must have mkfs.type)")
-	useGoCeph          = flag.Bool("go-ceph", false, "Use go-ceph library (default: false)")
-	useNbd             = flag.Bool("use-nbd", false, "Use rbd-nbd to map RBD Image (default: false)")
+	defaultImageFSType = flag.String("fs", "xfs", "FS type for the created RBD Image (must be xfs now)")
+	useGoCeph          = flag.Bool("go-ceph", false, "Use go-ceph library")
+	useNbd             = flag.Bool("use-nbd", true, "Use rbd-nbd to map RBD Image")
 )
 
 // setup a validating flag for remove action
@@ -160,6 +161,7 @@ func main() {
 	if err != nil {
 		log.Printf("ERROR: Unable to create UNIX socket: %v", err)
 	}
+
 }
 
 // isDebugEnabled checks for RBD_DOCKER_PLUGIN_DEBUG environment variable