+++ CHANGELOG.md
@@ -8,6 +8,14 @@ We attempt to adhere to [Semantic Versioning](http://semver.org/).
 ### Removed
 ### Changed
 
+## [1.5.3] - 2017-04-26
+### Added
+- 'make rpms' can generate rpm packages
+- image lock preempt function
+
+### Removed
+- image auto rm
+
 ## [1.5.2] - 2017-03-29
 ### Added
 - User can map rbd image through rbd-nbd command now